refactor: move git package to gitops package
