+++ internal/entrypoints/mcpserver/tools/test_helpers_test.go
@@ -6,7 +6,7 @@ import (
 
 	"github.com/swarm-deploy/swarm-deploy/internal/differ"
 	"github.com/swarm-deploy/swarm-deploy/internal/differ/diff"
-	gitx "github.com/swarm-deploy/swarm-deploy/internal/git"
+	gitx "github.com/swarm-deploy/swarm-deploy/internal/gitops/git"
 	"github.com/swarm-deploy/swarm-deploy/internal/registry"
 	"github.com/swarm-deploy/swarm-deploy/internal/service"
 	"github.com/swarm-deploy/swarm-deploy/internal/swarm"
