+++ internal/gitops/git/lazy.go
