+++ internal/gitops/git/mock.go
